Update Netrw plugin to v165b
Previous version was v156.

This solves a problem that :Lexplore does not toggle (close) the Netrw
window when entering a directory or doing other edits in the Netrw
buffer. This problem is discussed here:
https://vi.stackexchange.com/questions/10988/toggle-explorer-window

Plugin source: http://www.drchip.org/astronaut/vim/index.html#NETRW

To install, download the vimball archive (netrw.vba) and install it by
opening the .vba file with Vim and running :source %.

This installs the plugin under the first entry of ':set runtimepath?',
which by default seems to be ~/.vim. The original version of the plugin
seems to still be there (e.g. /usr/local/Cellar/vim/8.0.0430/share/vim/vim80/plugin)
after installatin of the new version, but it seems to be overwritten by
the new version (i.e. the newly installed version in ~/.vim takes
effect).