Add commands for enabling bash-completion to .bashrc on Linux
This, among others, causes all scripts in /etc/bash_completion.d to be
sourced.

These commands are also in /etc/bash.bashrc, however, they seem to be
outcommented by default, e.g. on Ubuntu.

The added code also includes a check so that the commands are only
executed if bash-completion is not active yet, which is tested by
checking whether the _init_completion function exists.