Move content of .bash_profile into .bashrc
There's no real need for the settings being split across these two
files. We want every shell to look the same, no matter whether it's a
login or non-login shell (which was in practice already the case since
.bash_profile used to only set exported variables).

This also makes things clearer when starting an interactive root shell
with 'sudo -s' when $HOME of the invoking user is kept but env_reset is
still on (see 'man sudoers), as it's the case on macOS. In this case,
'sudo -s' only sources ~/.bashrc but not ~/.bash_profile, thus settings
that are made only in .bash_profile wouldn't be available in this shell.

By having everything in .bashrc, such things are avoided.