Configure Git Credential Manager (GCM)
Configured with 'git credential-manager configure'. This adds the
added lines in this commit to .gitconfig and, after performing the
authentication through the web browser, saves an 'application password'
entry named 'git:https://github.com' in Keychain Access.

Note: the current configuration works for macOS and likely Linux. For
the case of WSL2 on Windows, it would probably need to be adapted. This
is left for future work.