Extend 'funcfile' to 'locate' function
The new 'locate' function works for all types of commands and can thus
determine the type of any command (functions, aliases, files, and shell
builtins). For functions and aliases, the file including the line number
of the corresponding definition is output. If a command has multiple
definitions (e.g. multiple executables with the same name, or an
executable and an alias), all of these definitions are output.