Add 'lbr' function for listing functions and aliases in a .bashrc.* file
This function lists all the function and alias names defined in a given
.bashrc.* file, along with the line number at which the definition
occurs.

The function also defines section headers of the form

   #=--------------...#
   # ...
   #---------------...#

which may be used to structure a .bashrc.* file. If such headers are
present, they are printed along with the function and alias names.

Next steps:

Another related function to be done is a function that takes a function
or alias name as an argument and prints the documentation comment
adjacent to it (possibly along with the function definitions, as it is,
for example, output by 'declare -f'). To do this, another marker, like
for example:

    #+
    # ...
    # ...
    function() {

should probably be introduced to clearly detect the documentation comment
to be printed by this function (it would probably also be possible to
just use all # lines directly preceding the function or alias definition
until encountering a line that does NOT start with #).