Change detection of alias definitions in 'cmd' function
Simply parse ~/.bashrc.* files for alias definitions, instead of
analysing the Bash traces [1]. The reason is that the former solution
only detected the FIRST definition of an alias. That means, if an alias
was defined multiple time (e.g. two alias definitions in the same file
or in different files), the wrong alias definition was returned (since
the later definition overrided the former one).

The current solution returns the correct alias definition if an alias is
defined multiple times.

1. The detection of alias definitions is restricted to the ~/.bashrc.*
   files (i.e. if an alias is defined in a different file, it is not
   detected).
2. The alias definitions in the .bashrc.* file must use a specific
   formatting in order to be detected (e.g. only whitespace between the
   beginning of the line and the alias definition).
3. If an alias that has a definition in a file is redefined in the shell,
   the definition in the shell is not detected and the function (wrongly)
   still returns the definition in the file.

[1] https://unix.stackexchange.com/a/483353/317243