Add .dotfiles.info directory with README and install
Eventually replacing the README and install script in the gist [1].

The install script has been changed as follows:

- Repository (Git directory) is .dotfiles.git instead of .dotfiles
- Existing dotfiles with the same name as repository dotfiles are
  not automatically backed up, but listed and the user is prompted
  to confirm to proceed or not

Note: with the .dotfiles.info directory, no separate gist for the
README and install script is necessary anymore. The drawback is
that the .dotfiles.info directory will be included in the user's
HOME directory, however, this shouldn't be a big problem, and it
fits in well with the .dotfiles.git directory containing the Git
repository.

[1] https://gist.github.com/weibeld/869f723063811e5088708a9386bf52bf