Factor out Vim and Homebrew to modules, introduce init part of modules
Some settings of a module have to be done before all other modules are
loaded to prevent a dependenc on the order in which the modules are
loaded.

For example, the Homebrew module adds 'brew' to the PATH, but other
modules might rely on 'brew' being in the PATH e.g. for determining
whether Homebrew is used for package management.

The init part of a module is loaded before any of the modules are
loaded, so it can be guaranteed that anything that is set in the init
part of a module is available to all the modules.